feat: rename TextButton component
